+++ backend/src/controllers/user/index.js
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { signUp } from './signUp';
+
+const userRouter = Router();
+
+userRouter.post("/signUp", signUp);
+
+export { userRouter };
